add process management
process management enables functionality like SDG to be added to a process registry, which we will maintain in ~/.local/share/instructlab/internal/process_registry.json

the processes are created via subprocess.Popen. The way these work in `ilab` is as follows:

each command will have two handlers, one which click calls from the CLI pkg. This handler will parse the arguments into pickleable units. `add_process` is then called with whatever imports are necessary to be built into the subprocess call, and all kwargs.

the second handler is what the subprocess target is. in the case of SDG, this handler creates the proper server and then calls generate_data. Having both handlers in `InstructLab` allows us to manipulate the arguments easily

Signed-off-by: Charlie Doern <cdoern@redhat.com>